Reverting to php74 versions temporarily
Add back php74-openssl and php74-json and fix php74 socket links

Add in syslog-ng, enabled when passed an IP

Problem is runaway syslogd processes on newly upgraded 13.1 host,
where this image is the only image using syslogd and everything
else is syslog-ng. This is just a test to see if any change.