Version bump to rebuild
This is a reversal of the consul DNS setup. DNS resolution fails in cloned jails because there no 127.0.0.1 for some programs. It works if /etc/resolv.conf is updated to "nameserver cloned-ip-of-jail" but then consul DNS doesn't work again. Files will be left in place for future development. This can be added in as an option for VNET jails.