Version bump for rebuild
This is a maintenance update to include steps from FreeBSD wiki which don't work, yet are included for completeness now.

https://wiki.freebsd.org/Ports/net-im/mastodon has two non-working steps:

```
bundle config build.cbor --with-cflags="-Wno-incompatible-function-pointer-types" (Needed for 13.2-STABLE and 14.0-RELEASE and later)
bundle config build.posix-spawn --with-cflags="-Wno-incompatible-function-pointer-types" (Needed for 13.2-STABLE and 14.0-RELEASE and later)
```

Mastodon works so it's probably not necessary to include these, or maybe the cbor and posix-spawn gemfiles need install first.