New version with self-signed SSL certificate
This is an experimental version. It's not clear if it will work yet. It will create a self-signed certificate for nginx, and will talk to minio self-signed too.

It is expected this will have a secured frontend proxy like OPNSense haproxy, or another haproxy with valid certificate and public IP, for end-to-end HTTPS to minio.